# doc/api.yaml
openapi: 3.0.2
info:
  title: WAI-DEMO
  description: "Auth Client ID: 99w2F1wVLZq8GqJwZph1kE42GuAZFvlF"
  version: 1.0.0
servers:
- url: /api
paths:
...
  /form/:
    get:
      tags:
      - Form
      summary: get all forms
      operationId: get_form__get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Form  Get
                type: array
                items:
                  $ref: '#/components/schemas/FormDict'
    post:
      tags:
      - Form
      summary: add new form
      operationId: form_create_form__post
      parameters:
      - name: name
        in: query
        required: true
        style: form
        explode: true
        schema:
          title: Name
          type: string
      - name: description
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Description
          type: string
      - name: default_language
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Default Language
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_form_create_form__post'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormDict'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /form/{id}:
    get:
      tags:
      - Form
      summary: get form by id
      operationId: form_get_by_id_form__id__get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          title: Id
          type: integer
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormBase'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /webform/{id}:
    get:
      tags:
      - Form
      summary: get form by id
      operationId: webform_get_by_id_webform__id__get
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          title: Id
          type: integer
      - name: edit
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Edit
          type: boolean
          default: false
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    put:
      tags:
      - Form
      summary: update webform editor definition
      operationId: webform_update_webform__id__put
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          title: Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              title: Payload
              type: object
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /webform/:
    post:
      tags:
      - Form
      summary: post webform editor JSON value
      operationId: webform_create_webform__post
      requestBody:
        content:
          application/json:
            schema:
              title: Payload
              type: object
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema: {}
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /question/:
    get:
      tags:
      - Form
      summary: get all questions
      operationId: get_question__get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Question  Get
                type: array
                items:
                  $ref: '#/components/schemas/QuestionDict'
    post:
      tags:
      - Form
      summary: add new question
      operationId: question_create_question__post
      parameters:
      - name: name
        in: query
        description: question text
        required: false
        style: form
        explode: true
        schema:
          title: Name
          type: string
          description: question text
      - name: order
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Order
          type: integer
      - name: type
        in: query
        description: question type
        required: false
        style: form
        explode: true
        schema:
          $ref: '#/components/schemas/type'
      - name: meta
        in: query
        description: Wether question is metadata or not
        required: false
        style: form
        explode: true
        schema:
          title: Meta
          type: boolean
          description: Wether question is metadata or not
          default: false
      - name: required
        in: query
        description: Wether question is required or not
        required: false
        style: form
        explode: true
        schema:
          title: Required
          type: boolean
          description: Wether question is required or not
          default: true
      - name: min
        in: query
        description: "Minimum number for number question type, default: 0"
        required: false
        style: form
        explode: true
        schema:
          title: Min
          type: integer
          description: "Minimum number for number question type, default: 0"
          default: 0
      - name: max
        in: query
        description: Max number for number question type
        required: false
        style: form
        explode: true
        schema:
          title: Max
          type: integer
          description: Max number for number question type
      - name: form
        in: query
        description: "Existing form id, create if you don't have one"
        required: false
        style: form
        explode: true
        schema:
          title: Form
          type: integer
          description: "Existing form id, create if you don't have one"
      - name: question_group
        in: query
        description: "Name of Question Group, Append or Create"
        required: false
        style: form
        explode: true
        schema:
          title: Question Group
          type: string
          description: "Name of Question Group, Append or Create"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Body_question_create_question__post'
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuestionBase'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /option/:
    get:
      tags:
      - Option
      summary: get all options
      operationId: option_get_option__get
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Option Get Option  Get
                type: array
                items:
                  $ref: '#/components/schemas/OptionBaseWithId'
    post:
      tags:
      - Option
      summary: Add new option
      operationId: option_post_option__post
      parameters:
      - name: question_id
        in: query
        required: true
        style: form
        explode: true
        schema:
          title: Question Id
          type: integer
      - name: name
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Name
          type: string
      - name: order
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Order
          type: string
      - name: color
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Color
          type: string
      - name: score
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Score
          type: string
      - name: code
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Code
          type: string
      requestBody:
        content:
          application/json:
            schema:
              title: Translations
              type: array
              items:
                type: object
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OptionBaseWithId'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
  /data/form/{form_id}:
    get:
      tags:
      - Data
      summary: get all datas
      operationId: data_get_data_form__form_id__get
      parameters:
      - name: form_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          title: Form Id
          type: integer
      - name: page
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Page
          type: integer
          default: 1
      - name: perpage
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Perpage
          type: integer
          default: 10
      - name: administration
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Administration
          type: integer
      - name: question
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Question
          type: array
          items:
            type: integer
      - name: q
        in: query
        required: false
        style: form
        explode: true
        schema:
          title: Q
          type: array
          items:
            type: string
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataResponse'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
    post:
      tags:
      - Data
      summary: add new data
      operationId: data_create_data_form__form_id__post
      parameters:
      - name: form_id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          title: Form Id
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              title: Answers
              type: array
              items:
                $ref: '#/components/schemas/AnswerDict'
        required: true
      responses:
        "200":
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DataDict'
        "422":
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    Body_form_create_form__post:
      title: Body_form_create_form__post
      type: object
      properties:
        languages:
          title: Languages
          type: array
          items:
            type: string
        translations:
          title: Translations
          type: array
          items:
            type: object
...
    DataResponse:
      title: DataResponse
      required:
      - current
      - data
      - total
      - total_page
      type: object
      properties:
        current:
          title: Current
          type: integer
        data:
          title: Data
          type: array
          items:
            $ref: '#/components/schemas/DataDictWithHistory'
        total:
          title: Total
          type: integer
        total_page:
          title: Total Page
          type: integer
    FormBase:
      title: FormBase
      required:
      - id
      - name
      - question_group
      type: object
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        default_language:
          title: Default Language
          type: string
        languages:
          title: Languages
          type: array
          items:
            type: string
        translations:
          title: Translations
          type: array
          items:
            type: object
        question_group:
          title: Question Group
          type: array
          items:
            $ref: '#/components/schemas/QuestionGroupBase'
    FormDict:
      title: FormDict
      required:
      - default_language
      - description
      - id
      - languages
      - name
      - translations
      type: object
      properties:
        id:
          title: Id
          type: integer
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        default_language:
          title: Default Language
          type: string
        languages:
          title: Languages
          type: array
          items:
            type: string
        translations:
          title: Translations
          type: array
          items:
            type: object
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    type:
      description: question type
      allOf:
      - $ref: '#/components/schemas/QuestionType'
      default: text
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer